Rename `localizations` to `localization` and move the main script
