+++ src/localization/ru.js
