+++ src/localization/uz.js
